move ID unique check to link plugin
